refactor(metrics): extract langres.metrics package from core (shims + graph)
Move core/{metrics,analysis,debugging,diagnostics}.py to the new
langres.metrics package (git mv, rename-detected). Metrics *score* a
resolution -- they are not the modelling contract -- so they sit beside
langres.core, not in it. Back-compat shims (# TEMPORARY: deleted by the
W2 sweep) keep the old langres.core.* import paths working.

Repoint intra-package imports to siblings; repoint core/reports.py's
analysis/diagnostics imports to langres.metrics.* -- the reports->analysis
edge would otherwise GROW the {analysis, reports, plotting.blockers} knot
to 4. Measured: runtime SCC still [], all-edges [9,3,2,2,2]/18, the knot
RELOCATED to {metrics.analysis, core.reports, plotting.blockers}.

Add the langres.metrics prefix to the counterfactual mapping (keeps it
total) and swap the tangle baseline's one relocated member.
