+++ tools/refactor_target_packages.json
@@ -73,6 +73,7 @@
     "langres.core.trackers": "tracking",
     "langres.data": "data",
     "langres.data.data_profile": "data",
+    "langres.metrics": "metrics",
     "langres.plotting": "plotting",
     "langres.report": "report"
   },
